Watch newly created files
In a perfect world, we could simply run:

    gulp.watch(item.match, item.tasks);

Unfortunately, this would not "catch" new files. So we're using a plugin that does. Which means in order to retain this functionality, we also need something that replaces the way `gulp.run` or `gulp.start` used to work (as neither are still supported).

Hopefully we can simplify again when/if Gulp 4 is released.